Initial commit for setup.sh
Prepares Ubuntu Desktop for pentesting by loading many tools and program environments.  Burpsuite and some other custom tools may still need to be manually installed on a per case basis.  It also customizes the directory structure.  Read the usage at the start of the file for arguments to control processing.